Fix (workaround) dtype incompatibility between tf.data.Dataset and jax
Can't construct a tf.data.Dataset from output of jnp.array(...), because it's not one of the dtype specs.
Can't add any custom dtype specs outside of TensorFlow.core.
The low-level code in Praneeth's framework is -compiled, which throws an error if it encounters s.
Work-around: transform the contents of each batch as