Control per round over num_batches and max_epochs
User can now specify a number of batches to used for local SGD for each round
of communication, while simultaneously capping the number of epochs this
corresponds to. This can be especially relevant for datasets of vastly
different sizes. After each communication round, training resumes where it
last stopped, thus the entire dataset will be evenly explored (by increments
of a few batches).